Moved the Grid Square Gateway into the UI layer.
